requirements.txt
absl-py==2.2.1
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.5.1
bitsandbytes==0.45.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
datasets==3.5.0
Deprecated==1.2.18
diffusers==0.29.2
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
inquirerpy==0.3.4
Jinja2==3.1.6
lightning==2.5.1
lightning-utilities==0.14.3
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
peft==0.15.2
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
rapidata==2.20.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentry-sdk==2.24.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tinytag==2.1.0
tokenizers==0.21.1
torch==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
wandb==0.19.8
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
